Made hosts updates atomic, started adding/removing routes for all IPs of all remotes, instead of just the current IP.

+++ vpnfailsafe.sh
@@ -2,8 +2,6 @@
 
 set -eEuo pipefail
 
-readonly prog="$(basename "$0")"
-readonly private_nets="127.0.0.0/8,10.0.0.0/8,172.16.0.0/12,192.168.0.0/16"
 readonly dev
 readonly ${!foreign_option_*}
 readonly ifconfig_local
@@ -16,6 +14,10 @@ readonly route_vpn_gateway
 readonly script_type
 readonly trusted_ip
 
+readonly prog="$(basename "$0")"
+readonly private_nets="127.0.0.0/8,10.0.0.0/8,172.16.0.0/12,192.168.0.0/16"
+readonly -a remotes=($(env|grep -oP 'remote_[0-9]+=.*'|sort -g|cut -d= -f2))
+
 # $@ := ""
 error() {
     exit 1
@@ -23,20 +25,19 @@ error() {
 
 # $@ := "up" | "down"
 update_hosts() {
-    pkill -P1 -f "$prog" || true
+    pkill -P1 -f "bash $0" || true
     if [[ "$@" == up ]]; then
         uh() {
             local -r beg="# VPNFAILSAFE BEGIN" end="# VPNFAILSAFE END"
-            local -ar remotes=($(env|grep -oP 'remote_[0-9]+=.*'|cut -d= -f2))
-            local -r maybe_inplace="$([[ "$PPID" -eq 1 ]] && echo -i || echo)"
-            {
-                sed $maybe_inplace /etc/hosts -e "/^$beg/,/^$end/d" 
-                echo "$beg"
-                getent hosts "${remotes[@]}"|grep -v :
-                echo "$end"
-            } >/etc/hosts.vpnfailsafe
-            chmod --reference=/etc/hosts /etc/hosts.vpnfailsafe
-            mv /etc/hosts.vpnfailsafe /etc/hosts
+            if remote_entries="$(getent -s dns hosts "${remotes[@]}"|grep -v :)"; then
+                {
+                    sed -e "/^$beg/,/^$end/d" /etc/hosts
+                    echo -e "$beg\n$remote_entries\n$end"
+                } >/etc/hosts.vpnfailsafe
+                chmod --reference=/etc/hosts /etc/hosts.vpnfailsafe
+                sync /etc/hosts.vpnfailsafe
+                mv /etc/hosts.vpnfailsafe /etc/hosts
+            fi
         }
         uh
         while true; do sleep 5m; uh; done& 
@@ -45,15 +46,18 @@ update_hosts() {
 
 # $@ := "up" | "down"
 update_routes() {
+    remote_ips="$(getent -s files hosts "${remotes[@]}"|cut -d' ' -f1)"
     if [[ "$@" == up ]]; then
-        if [[ -z $(ip route show "$trusted_ip") ]]; then
-            ip route add "$trusted_ip/32" via "$route_net_gateway"
-        fi
+        for remote_ip in $remote_ips; do
+            if [[ -z $(ip route show "$remote_ip") ]]; then
+                ip route add "$remote_ip" via "$route_net_gateway"
+            fi
+        done
         for net in 0.0.0.0/1 128.0.0.0/1; do
             [[ -z $(ip route show "$net") ]] && ip route add "$net" via "$route_vpn_gateway"
         done
     elif [[ "$@" == down ]]; then
-        for route in "$trusted_ip/32" 0.0.0.0/1 128.0.0.0/1; do
+        for route in $remote_ips 0.0.0.0/1 128.0.0.0/1; do
             [[ -n $(ip route show "$route") ]] && ip route del "$route"
         done
         if [[ -n $(ip addr show dev "$dev" 2>/dev/null) ]]; then
@@ -83,8 +87,6 @@ update_resolv() {
 
 # $@ := ""
 update_firewall() {
-    local -r public_nic="$(ip route show "$trusted_ip"|cut -d' ' -f5)"
-
     # $@ := "INPUT" | "OUTPUT" | "FORWARD"
     insert_chain() {
         if iptables -C "$*" -j "VPNFAILSAFE_$*" 2>/dev/null; then
@@ -104,13 +106,13 @@ update_firewall() {
             OUTPUT) local -r sd=d states=NEW, io=o;;
             *) error;;
         esac
-        for remote in ${!remote_*}; do
-            if [[ "$remote" =~ ^remote_[0-9]+$ ]]; then
-                local port="remote_port_${remote##*_}"
-                local proto="proto_${remote##*_}"
-                iptables -A "VPNFAILSAFE_$*" -p "${!proto}" -"$sd" "${!remote}" --"$sd"port "${!port}" \
-                    -m conntrack --ctstate "$states"RELATED,ESTABLISHED -"$io" "$public_nic" -j ACCEPT
-            fi
+        local -r public_nic="$(ip route show "$trusted_ip"|cut -d' ' -f5)"
+        local i=1; for remote in "${remotes[@]}"; do
+            local port="remote_port_$i"
+            local proto="proto_$i"
+            iptables -A "VPNFAILSAFE_$*" -p "${!proto}" -"$sd" "$remote" --"$sd"port "${!port}" \
+                -m conntrack --ctstate "$states"RELATED,ESTABLISHED -"$io" "${public_nic:?}" -j ACCEPT
+            i=$((i + 1))
         done
     }
